move RegisterNode to package justin.db.actors.protocol
