+++ justin-core/src/main/scala/justin/db/actors/ClusterSubscriberActor.scala
@@ -5,7 +5,7 @@ import akka.cluster.ClusterEvent.{CurrentClusterState, MemberUp}
 import akka.cluster.{Cluster, MemberStatus}
 import justin.consistent_hashing.{NodeId, Ring}
 import justin.db.ClusterMembers
-import justin.db.actors.ClusterSubscriberActorProtocol.RegisterNode
+import justin.db.actors.protocol.RegisterNode
 
 trait ClusterSubscriberActor { self: Actor =>
 
@@ -31,8 +31,4 @@ trait ClusterSubscriberActor { self: Actor =>
       nodeRef        = context.actorSelection(RootActorPath(address) / "user" / nodeName)
     } yield nodeRef ! RegisterNode(nodeId)
   }
-}
-
-object ClusterSubscriberActorProtocol {
-  case class RegisterNode(nodeId: NodeId)
 }
