+++ justin-core/src/main/scala/justin/db/actors/protocol/RegisterNode.scala
@@ -1,0 +1,5 @@
+package justin.db.actors.protocol
+
+import justin.consistent_hashing.NodeId
+
+case class RegisterNode(nodeId: NodeId)
