+++ justin-core/src/test/scala/justin/db/actors/ClusterSubscriberActorTest.scala
@@ -4,7 +4,7 @@ import akka.actor.{Actor, ActorSystem}
 import akka.testkit.{ImplicitSender, TestActorRef, TestKit}
 import com.typesafe.config.ConfigFactory
 import justin.consistent_hashing.{NodeId, Ring}
-import justin.db.actors.ClusterSubscriberActorProtocol.RegisterNode
+import justin.db.actors.protocol.RegisterNode
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{BeforeAndAfterAll, FlatSpecLike, Matchers}
